requirements.txt
telethon>=1.35.0
python-telegram-bot>=20.7
gspread>=5.12.0
google-auth-oauthlib>=0.7.0
google-api-python-client>=2.88.0
openai>=1.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
Flask
gunicorn
requests